Rename ichnaea/config.py to ichnaea/conf.py
The words "config" and "configuration" are used in different contexts
which makes this project harder to work on.

This renames the "config.py" file to "conf.py". This file manages
application settings derived from constants and the environment for
all the services. This is the first step of a settings overhaul.
The next step puts everything into a "settings" structure in the
conf module. This is a little like what Django does. It adds some
convenience for things like testing.
