+++ ichnaea/alembic/versions/000000000000_base.py
@@ -13,7 +13,7 @@ from alembic import op
 import sqlalchemy as sa
 from sqlalchemy import text
 
-from ichnaea.config import (
+from ichnaea.conf import (
     DB_RO_URI,
     DB_RW_URI,
 )
